pyproject.toml
[tool.poetry]
name = "reagent"
version = "1.0.0"
description = ""
authors = ["Maguire <papaymaguire@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
pydantic = "^2.10.6"
pydantic-settings = "^2.7.1"
httpx = "^0.28.1"
openai = "^1.63.0"
cryptography = "^44.0.2"
sqlmodel = "^0.0.24"
alembic = "^1.15.1"
hatchet-sdk = "^0.47.0"
fast-depends = "^2.4.12"
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
sse-starlette = "^2.2.1"
python-multipart = "^0.0.20"
typer = {version = "^0.15.2", optional = true}

[tool.poetry.extras]
cli=["typer"]

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
testcontainers = "^4.9.0"
mypy = "^1.13.0"
pytest-cov = "^6.0.0"
mkdocs-material = "^9.6.7"
psutil = "^7.0.0"
astor = "^0.8.1"


[tool.pytest.ini_options]
# Basic pytest configuration
testpaths = ["tests"]
python_files = "*_test.py"
# For pytest-asyncio
asyncio_mode = "auto"

[tool.coverage.run]
source = ["reagent"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:"
]

# Optional HTML report settings
[tool.coverage.html]
directory = "coverage_html_report"

[tool.pyright]
strict = true
reportMissingTypeStubs = true
reportUnboundVariable = true
reportMissingImports = false
reportOptionalMemberAccess = true
reportGeneralTypeIssues = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
